backend/requirements.txt
ajsonrpc==1.2.0
anyio==4.8.0
blinker==1.9.0
bottle==0.13.2
c3d==0.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distutils-pytest==0.2.1
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.56.0
freetype-py==2.5.1
fsspec==2025.3.0
h11==0.14.0
huggingface-hub==0.29.3
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
open3d-python==0.3.0.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformio==6.1.17
pluggy==1.5.0
pyelftools==0.32
pyglet==2.1.3
PyOpenGL==3.1.0
pyparsing==3.2.1
pyrender==0.1.45
pyserial==3.5
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
shapely==2.0.7
six==1.17.0
smplx==0.1.28
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.3
trimesh==4.6.5
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wsproto==1.2.0
